refactor(init): ♻️ add type hints and improve null safety
- added PHP type hints to controller methods
- ensured null safety with appropriate instanceof checks
- refactored controllers for better reliability and maintainability

+++ app/Http/Controllers/Approvals/ApprovalComponentContributorController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers\Approvals;
 
 use App\Http\Controllers\Controller;
 use App\Models\Approval\ApprovalContributor;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\Request;
+use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Facades\App;
 use Illuminate\Validation\Rule;
 
@@ -14,6 +16,9 @@ class ApprovalComponentContributorController extends Controller
      * Approval Component Contributor Index
      *
      * Display a listing of the resource.
+     */
+    /**
+     * @return ApprovalContributor[]|Collection<int, ApprovalContributor>|LengthAwarePaginator<int, ApprovalContributor>
      */
     public function index(Request $request)
     {
@@ -38,6 +43,9 @@ class ApprovalComponentContributorController extends Controller
      *
      * @noinspection DuplicatedCode
      */
+    /**
+     * @return array<string, string>
+     */
     public function store(Request $request)
     {
         $request->validate([
@@ -46,7 +54,7 @@ class ApprovalComponentContributorController extends Controller
         ]);
 
         $approvalComponentContributor = new ApprovalContributor;
-        $approvalComponentContributor->approval_component_id = (int)$request->route('approval_component_id');
+        $approvalComponentContributor->approval_component_id = (int) $request->route('approval_component_id');
         $approvalComponentContributor->approvable_id = $request->input('approvable_id');
         $approvalComponentContributor->approvable_type = config('approval.group')[$request->input('key')];
         $approvalComponentContributor->save();
@@ -60,6 +68,9 @@ class ApprovalComponentContributorController extends Controller
      * Approval Component Contributor Show
      *
      * Show the specified resource.
+     */
+    /**
+     * @return ApprovalContributor|Collection<int, ApprovalContributor>
      */
     public function show(Request $request)
     {
@@ -76,6 +87,9 @@ class ApprovalComponentContributorController extends Controller
      *
      * @noinspection DuplicatedCode
      */
+    /**
+     * @return array<string, string>
+     */
     public function update(Request $request)
     {
         $request->validate([
@@ -84,10 +98,12 @@ class ApprovalComponentContributorController extends Controller
         ]);
 
         $approvalComponentContributor = ApprovalContributor::findOrFail($request->route('id'));
-        $approvalComponentContributor->approval_component_id = (int)$request->route('approval_component_id');
-        $approvalComponentContributor->approvable_id = $request->input('approvable_id');
-        $approvalComponentContributor->approvable_type = config('approval.group')[$request->input('key')];
-        $approvalComponentContributor->save();
+        if ($approvalComponentContributor instanceof ApprovalContributor) {
+            $approvalComponentContributor->approval_component_id = (int) $request->route('approval_component_id');
+            $approvalComponentContributor->approvable_id = $request->input('approvable_id');
+            $approvalComponentContributor->approvable_type = config('approval.group')[$request->input('key')];
+            $approvalComponentContributor->save();
+        }
 
         return [
             'message' => trans('messages.success.update', ['target' => 'contributor'], App::getLocale()),
@@ -99,10 +115,15 @@ class ApprovalComponentContributorController extends Controller
      *
      * Remove the specified resource from storage.
      */
+    /**
+     * @return array<string, string>
+     */
     public function delete(Request $request)
     {
         $approvalComponentContributor = ApprovalContributor::findOrFail($request->route('id'));
-        $approvalComponentContributor->delete();
+        if ($approvalComponentContributor instanceof ApprovalContributor) {
+            $approvalComponentContributor->delete();
+        }
 
         return [
             'message' => trans('messages.success.delete', ['target' => 'contributor'], App::getLocale()),
